+++ cyvcf2/helpers.h
@@ -8,4 +8,5 @@ int gt_types_012_from_int8(const int8_t *data, int num_samples, int ploidy,
                            int32_t *gt_types, int32_t *gt_idxs,
                            int32_t *gt_phased);
 int vcf_line_strip_format(char *line, int len, const char *keep);
+int cyvcf2_hdr_set_parse_formats(bcf_hdr_t *hdr, const char *fmts);
 int32_t* bcf_hdr_seqlen(const bcf_hdr_t *hdr, int32_t *nseq);